build: warning-free clean build (fix a real C4244 in our code)
quantize() stored std::lround's long result in a float before casting to
int — a C4244 narrowing (and a pointless round-trip) introduced with the
24-bit position change; now stays in the integer domain. The rest was
third-party noise burying it: enet's C5287, stb's sprintf C4996, and the
deps' CMake dev/deprecation warnings are now silenced at their targets
(enet /w, StbImpl 4996, -Wno-dev on configure). MeatEngine/MeatTests keep
full /W4 -Wall -Wextra; a clean build now emits zero warnings.

+++ src/engine/net/DeltaSnapshot.cpp
@@ -22,10 +22,11 @@ constexpr std::int32_t kPosQMax = (1 << 23) - 1; // largest signed 24-bit magnit
 constexpr float kPosMax = static_cast<float>(kPosQMax) / kPosStepsPerMeter; // ±32767.996 m
 
 std::int32_t quantize(float v) {
-    const float scaled = std::lround(v * kPosStepsPerMeter);
+    const long scaled = std::lround(v * kPosStepsPerMeter); // integer domain: no float re-narrowing
     // A position past ±32 km is beyond any reachable voxel world; clamp so the wire stays valid,
     // but assert in dev builds so it surfaces as a bug rather than a silent teleport in release.
-    assert(scaled >= -kPosQMax && scaled <= kPosQMax && "snapshot position exceeds ±32 km wire range");
+    assert(scaled >= -static_cast<long>(kPosQMax) && scaled <= static_cast<long>(kPosQMax) &&
+           "snapshot position exceeds ±32 km wire range");
     return std::clamp(static_cast<std::int32_t>(scaled), -kPosQMax, kPosQMax);
 }
 float dequantize(std::int32_t q) { return static_cast<float>(q) / kPosStepsPerMeter; }
